refactor: move useButton to its own folder.
