+++ docs/data/base/components/button/button.md
@@ -99,7 +99,7 @@ Similarly, `<ButtonUnstyled component="span" type="reset">` will not reset its p
 ## Hook
 
 ```js
-import { useButton } from '@mui/base/ButtonUnstyled';
+import useButton from '@mui/base/useButton';
 ```
 
 The `useButton` hook lets you apply the functionality of `ButtonUnstyled` to a fully custom component.
