+++ packages/mui-base/src/ButtonUnstyled/ButtonUnstyled.types.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { OverrideProps, Simplify } from '@mui/types';
-import { UseButtonParameters, UseButtonRootSlotProps } from './useButton.types';
+import { UseButtonParameters, UseButtonRootSlotProps } from '@mui/base/useButton';
 import { SlotComponentProps } from '../utils';
 
 export interface ButtonUnstyledActions {
