+++ packages/mui-base/src/ButtonUnstyled/index.ts
@@ -4,5 +4,3 @@ export {
   getButtonUnstyledUtilityClass,
 } from './buttonUnstyledClasses';
 export * from './ButtonUnstyled.types';
-export { default as useButton } from './useButton';
-export * from './useButton.types';
