+++ packages/mui-base/src/useButton/index.ts
@@ -1,0 +1,5 @@
+export { default as useButton } from './useButton';
+export { default } from './useButton';
+export * from './useButton';
+
+export * from './useButton.types';
