+++ packages/mui-base/src/useMenuItem/useMenuItem.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { unstable_useId as useId, unstable_useForkRef as useForkRef } from '@mui/utils';
 import { MenuUnstyledContext } from '../MenuUnstyled';
-import { useButton } from '../ButtonUnstyled';
+import useButton from '@mui/base/useButton';
 import { UseMenuItemParameters } from './useMenuItem.types';
 
 export default function useMenuItem(props: UseMenuItemParameters) {
