+++ packages/mui-base/src/useSelect/useSelect.ts
@@ -3,7 +3,7 @@ import {
   unstable_useControlled as useControlled,
   unstable_useForkRef as useForkRef,
 } from '@mui/utils';
-import { useButton } from '../ButtonUnstyled';
+import useButton from '@mui/base/useButton';
 import {
   SelectOption,
   UseSelectButtonSlotProps,
