+++ packages/mui-base/src/useSelect/useSelect.types.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { UseButtonRootSlotProps } from '../ButtonUnstyled';
+import { UseButtonRootSlotProps } from '@mui/base/useButton';
 import {
   OptionState,
   UseListboxOptionSlotProps,
